Slight change...
Instead of creating another array and then running through that again, I changed it to just a simple comparison and only storing the lowest of the 9 potential options. Also removed the creation of the 'row' array, it wasn't needed except for readability but there is so little code I decided it doesn't make a huge difference but eh.